Move waiting lists models directory in line with other models
