+++ app/Models/Training/WaitingList.php
@@ -3,6 +3,8 @@
 namespace App\Models\Training;
 
 use App\Models\Mship\Account;
+use App\Models\Training\WaitingList\WaitingListAccount;
+use App\Models\Training\WaitingList\WaitingListFlag;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Database\Eloquent\SoftDeletes;
