+++ tests/Feature/Training/WaitingListFeatureTest.php
@@ -7,10 +7,9 @@ use App\Events\Training\AccountChangedStatusInWaitingList;
 use App\Events\Training\AccountDemotedInWaitingList;
 use App\Events\Training\AccountNoteChanged;
 use App\Events\Training\AccountPromotedInWaitingList;
-use App\Events\Training\AccountRemovedFromWaitingList;
 use App\Models\Mship\Account;
 use App\Models\Training\WaitingList;
-use App\Models\Training\WaitingListFlag;
+use App\Models\Training\WaitingList\WaitingListFlag;
 use App\Services\Training\AddToWaitingList;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Illuminate\Support\Facades\Event;
