+++ tests/Unit/Training/WaitingListAccountTest.php
@@ -4,8 +4,7 @@ namespace Tests\Unit\Training;
 
 use App\Models\Mship\Account;
 use App\Models\NetworkData\Atc;
-use App\Models\Training\WaitingListFlag;
-use App\Models\Training\WaitingListStatus;
+use App\Models\Training\WaitingList\WaitingListStatus;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Tests\TestCase;
