+++ tests/Unit/Training/WaitingListFlagTest.php
@@ -2,15 +2,15 @@
 
 namespace Tests\Unit\Training;
 
-use App\Models\Mship\Account;
-use App\Models\Atc\Endorsement;
-use App\Models\NetworkData\Atc;
-use App\Models\Training\WaitingList;
-use App\Models\Training\WaitingListFlag;
-use App\Models\Training\WaitingListStatus;
-use App\Services\Training\AddToWaitingList;
 use App\Events\Training\AccountAddedToWaitingList;
 use App\Listeners\Training\WaitingList\AssignFlags;
+use App\Models\Atc\Endorsement;
+use App\Models\Mship\Account;
+use App\Models\NetworkData\Atc;
+use App\Models\Training\WaitingList;
+use App\Models\Training\WaitingList\WaitingListFlag;
+use App\Models\Training\WaitingList\WaitingListStatus;
+use App\Services\Training\AddToWaitingList;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Tests\TestCase;
