+++ tests/Unit/Training/WaitingListTest.php
@@ -4,7 +4,7 @@ namespace Tests\Unit\Training;
 
 use App\Models\Mship\Account;
 use App\Models\Training\WaitingList;
-use App\Models\Training\WaitingListFlag;
+use App\Models\Training\WaitingList\WaitingListFlag;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Tests\TestCase;
